feat: refactor address management by consolidating address components and enhancing address update functionality
